fix for Git issue 11235 - Kafka provider did not terminate on thrown (#60)
* fix for Git issue 11235 - Kafka provider did not terminate on thrown
exception during startup. ( Connection error to bmwhiskxxx trigger
database)

* fix trailing blanks

* leave memory_leak check as Thread, because only a thread shares the
memory whit the main process ( staring the memory_leak as subprocess has
no access on the memory of the main process). 
Move the memory leak thread start to the end of main,  that it will not
prevent the error handling in the startup of the kafka provider

* fix trailing blanks

Co-authored-by: 062088724 <062088724@DESKTOP-QA6FT6P>